Add configurable bitdepth support to the Flex driver
Necessary refactoring: swap the input stream reading and client command
forwarding coroutine nesting in order to be able to reset the input
stream upon bitdepth change.

Restarting the reader is a both a clean approach to reseting the
protocol parsing state machine and also solves a particular issue of the
reader hanging on insufficient input in the following scenario:

- Driver is running at 8bit depth
- Reader is midway reading a frame
- Driver/Firmware is updated to run at 12bit depth
- Reader now expects the next sample to be 12 bits, but the current
  buffered frame still contains an 8bit sample.
- Reader is now stuck on `reader.ReadByte()` forever

These changes are intended to be backwards compatible with older
versions of Play (which send no bitdepth commands and assume that the
driver sets 8bit mode by default).

Also add some notes on different behaviours between the generations.