Assume default 12bit bitdepth in firmware and client
Since there is no "set bitdepth 12" command available in the Flex
firmware, we remove the default command of UL (set bitdepth=8) from
the driver and assume a default of 12bit (which is the case with
at least v4 and v5 firmware).

This makes the driver compatible only with v5 firmware, since v4
firmware does not seem to handle 12bit properly.

It is assumed that the client (Play) does _not_ send any bitdepth
commands.