Add support Flex device info, refine auto-connect, add broadcasts
This extends the Driver to:
- Support Flex device metadata in Driver message replies
- Enable broadcasting messages to all connected clients

Using it, the Flex backend implements a flow where it auto-connects to
devices and informs about device changes by broadcasting a Status
update. This eliminates the need for the client to periodically poll the
Driver.

The protocol changes are backwards compatible for the /senso endpoint.